move ripple button and write tests
